# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.AuthorForList:
    properties:
      avatar_url:
        type: string
      birth_day:
        type: string
      book_count:
        type: integer
      died_year:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
  models.AuthorReq:
    properties:
      about_text:
        type: string
      avatar_url:
        type: string
      birth_day:
        example: "1960-02-15"
        type: string
      country:
        type: string
      creativity:
        type: string
      died_year:
        example: "2000-03-24"
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
  models.AuthorRes:
    properties:
      about_text:
        type: string
      avatar_url:
        type: string
      birth_day:
        example: "1960-02-15"
        type: string
      books:
        items:
          $ref: '#/definitions/models.BooksForList'
        type: array
      country:
        type: string
      created_at:
        type: string
      creativity:
        type: string
      died_year:
        example: "2000-03-24"
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
      updated_at:
        type: string
    type: object
  models.AuthorUpdateReq:
    properties:
      about_text:
        type: string
      avatar_url:
        type: string
      birth_day:
        example: "1960-02-15"
        type: string
      country:
        type: string
      creativity:
        type: string
      died_year:
        example: "2000-03-24"
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
    type: object
  models.Authors:
    properties:
      _meta:
        $ref: '#/definitions/models.Meta'
      authors:
        items:
          $ref: '#/definitions/models.AuthorForList'
        type: array
    type: object
...
  models.BooksForList:
    properties:
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      like_count:
        type: integer
      sub_category_id:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
...
  models.Meta:
    properties:
      current_page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Book store API
  version: "1.0"
paths:
...
  /author:
    post:
      consumes:
      - application/json
      description: Here author can update info
      operationId: createauthor
      parameters:
      - description: AuthorCreate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthorReq'
      produces:
      - application/json
      responses:
        "201":
          description: Author data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthorRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: create author
      tags:
      - Author
    put:
      consumes:
      - application/json
      description: Here author can update info
      operationId: updateauthor
      parameters:
      - description: UpdateAuthor
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthorUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Author data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthorRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: put author
      tags:
      - Author
  /author/{id}:
    delete:
      consumes:
      - application/json
      description: Here author can DELETE info
      operationId: deleteauthor
      parameters:
      - description: DELETEAuthor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Author delete
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: DELETE author
      tags:
      - Author
    get:
      consumes:
      - application/json
      description: Here author can get info
      operationId: getauthor
      parameters:
      - description: GetAuthor
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Author data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthorRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET author
      tags:
      - Author
  /author/list:
    get:
      consumes:
      - application/json
      description: Here author can get info
      operationId: getauthors
      parameters:
      - description: LIMIT
        in: query
        name: limit
        type: integer
      - description: PAGE_NUMBER
        in: query
        name: page
        type: integer
      - description: SEARCH
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Author data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Authors'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET author
      tags:
      - Author
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"